Rename default FlightClientConfigs to FlightClientStreamConfig
Since this is more about specifically the streaming protocol and I'll add
other dimensions that don't map 1:1.
