+++ packages/react-client/src/ReactFlightClientStreamConfigWeb.js
