+++ packages/react-client/src/forks/ReactFlightClientConfig.dom-fb-experimental.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-export * from 'react-client/src/ReactFlightClientConfigBrowser';
+export * from 'react-client/src/ReactFlightClientStreamConfigWeb';
 export * from 'react-dom-bindings/src/shared/ReactFlightClientConfigDOM';
 export * from 'react-server-dom-fb/src/ReactFlightClientConfigFBBundler';
